Refactor formatting helpers
